Refactor branch response parsing into `parse-branch`
