--- test/unit/__fixtures__/create-fixture.js
@@ -1,6 +1,0 @@
-
-module.exports.createFixture = (stdOut, stdErr) => ({
-   stdOut,
-   stdErr,
-   parserArgs: [stdOut, stdErr],
-});
