+++ test/unit/__fixtures__/pull/pull-with-merge-conflicts.js
@@ -1,4 +1,4 @@
-const {createFixture} = require('../create-fixture');
+const {createFixture} = require('../../../helpers');
 
 const stdErr = ``;
 const stdOut = ``;
