+++ test/unit/__fixtures__/push/push-deleted-branch.js
@@ -1,4 +1,4 @@
-const {createFixture} = require('../create-fixture');
+const {createFixture} = require('../../../helpers');
 
 const stdOut = `
 To github.com:kwsites/mock-repo.git
