+++ test/unit/__fixtures__/push/push-new-branch-remote-says-vulnerabilities.js
@@ -1,4 +1,4 @@
-const {createFixture} = require('../create-fixture');
+const {createFixture} = require('../../../helpers');
 const {gitHubPullRequest, gitHubAlertsUrl} = require('./constants');
 
 const stdErr = `Pushing to git@github.com:kwsites/mock-repo.git
