+++ test/unit/__fixtures__/push/push-new-branch.js
@@ -1,4 +1,4 @@
-const {createFixture} = require('../create-fixture');
+const {createFixture} = require('../../../helpers');
 const {gitLabPullRequest} = require('./constants');
 
 const stdOut = `To github.com:kwsites/mock-repo.git
